Moved most of stindex code into lib.
